requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.1.2
Brotli==1.0.9
cachetools==4.2.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap3==2022.1
crispy-bootstrap4==2022.1
cryptography==41.0.7
cssselect2==0.7.0
Django==4.2.2
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-extensions==3.2.3
django-googledrive-storage==1.6.0
django-phonenumber-field==7.1.0
django-scheduler==0.10.1
django-storages==1.13.2
fonttools==4.40.0
frozenlist==1.3.3
google-api-core==2.10.2
google-api-python-client==2.90.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-oauth==1.0.1
googleapis-common-protos==1.59.1
gunicorn==20.1.0
html5lib==1.1
httplib2==0.22.0
icalendar==5.0.11
idna==3.4
mssql-django==1.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
paramiko==3.4.0
phonenumbers==8.13.14
Pillow==9.5.0
protobuf==4.23.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydyf==0.6.0
pygsheets==2.0.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyodbc==5.1.0
pyOpenSSL==23.3.0
pyparsing==3.1.0
pyphen==0.14.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.39.2
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
tinycss2==1.2.1
tqdm==4.65.0
twilio==8.3.0
typing_extensions==4.6.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.16
weasyprint==59.0
webencodings==0.5.1
yarl==1.9.2
zopfli==0.2.2
